Remove git/object.rs and create sub-modules
