+++ src/vcs/git/stats.rs
@@ -1,0 +1,36 @@
+// This file is part of radicle-surf
+// <https://github.com/radicle-dev/radicle-surf>
+//
+// Copyright (C) 2019-2020 The Radicle Team <dev@radicle.xyz>
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License version 3 or
+// later as published by the Free Software Foundation.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+pub use git2::Oid;
+
+#[cfg(feature = "serialize")]
+use serde::Serialize;
+
+/// Stats for a repository
+#[cfg_attr(
+    feature = "serialize",
+    derive(Serialize),
+    serde(rename_all = "camelCase")
+)]
+pub struct Stats {
+    /// Number of commits
+    pub commits: usize,
+    /// Number of local branches
+    pub branches: usize,
+    /// Number of contributors
+    pub contributors: usize,
+}
